requirements.txt
Flask
pandas
numpy
scikit-learn
xgboost
pefile
google-generativeai
joblib
matplotlib
plotly
watchdog
flask-socketio
python-socketio
